Deep Populate To Show Friend Users Of Current User
The current user has a virtual field called friends, which is accessed
via populate. This friends field contains a field called friendUsers
which are ObjectIds of User documents. This is accessed via populate.
In order to obtain this second layer of User documents, it is necessary
to use a specific syntax involving the path property, and nesting
populate calls.